e presccription api changes

+++ EasyHMSAPI.Application/Handlers/CommandHandlers/DeletePersonalizedDataHandler.cs
@@ -16,10 +16,27 @@ namespace EasyHMSAPI.Application.Handlers.CommandHandlers
 
         public async Task<DeletePersonalizedDataResponseModel> Handle(DeletePersonalizedDataRequestModel request, CancellationToken cancellationToken)
         {
-            var existing = await _dbContext.LookupPersonals.FirstOrDefaultAsync(lp => lp.PersonalId == request.PersonalId && lp.DoctorID == request.DoctorId, cancellationToken);
+            var existingDoctor = await _dbContext.Doctors
+               .Where(x => x.DoctorID == request.DoctorId)
+               .FirstOrDefaultAsync(cancellationToken);
+            if (existingDoctor == null)
+            {
+                return new DeletePersonalizedDataResponseModel { Message = "Doctor not found." };
+            }
+
+            var existingHospital = await _dbContext.Hospitals
+                .Where(x => x.HospitalID == request.HospitalId)
+                .FirstOrDefaultAsync(cancellationToken);
+            if (existingHospital == null)
+            {
+                return new DeletePersonalizedDataResponseModel { Message = "Hospital not found." };
+            }
+
+            var existing = await _dbContext.LookupPersonals
+                .FirstOrDefaultAsync(lp => lp.PersonalId == request.PersonalId && lp.DoctorID == request.DoctorId, cancellationToken);
             if (existing == null)
             {
-                return new DeletePersonalizedDataResponseModel { Message = "NotFound" };
+                return new DeletePersonalizedDataResponseModel { Message = "Not Found" };
             }
 
             _dbContext.LookupPersonals.Remove(existing);
